Modify loading of the models
