+++ genai-engine/.env
@@ -19,6 +19,7 @@ GENAI_ENGINE_ENABLE_PERSISTENCE=enabled
 ######################
 # [local] or anything else for remote
 GENAI_ENGINE_ENVIRONMENT=local
+GENAI_ENGINE_LOG_FORMAT='[%(asctime)s] [%(process)d] [%(levelname)s] [%(module)s] %(message)s'
 
 ##################################
 #### LLM for Rule Evaluations ####
