+++ genai-engine/src/docker-entrypoint.sh
@@ -47,4 +47,4 @@ echo "==> Running database migration"
 poetry run alembic upgrade head || exit 1
 
 echo "==> Starting the GenAI Engine server with ${WORKER:-1} worker"
-poetry run gunicorn src.server:get_app -c /app/src/gunicorn.conf.py --preload
+poetry run gunicorn src.server:get_app -c /app/src/gunicorn.conf.py
